Improve custom listing fields admin quality
